emailer/requirements.txt
Flask==3.0.3
Flask-Mail
python-dotenv